+++ @navikt/core/react/src/util/hooks/Hooks.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import { useScrollLock } from "./useScrollLock";
 
 export default {
@@ -48,25 +48,70 @@ export const UseScrollLockScrollXAndY: StoryObj = {
 export const UseScrollLockResizeTest: StoryObj = {
   render: ScrollLockComponent,
   decorators: [
+    (story) => {
+      const [width, setWidth] = useState(400);
+      return (
+        <div
+          style={{
+            background: "var(--ax-bg-softA)",
+          }}
+        >
+          <div>{story()}</div>
+          <button onClick={() => setWidth(() => 80)}>Scroll</button>
+          <button onClick={() => setWidth(() => 400)}>No scroll</button>
+          <div style={{ width }}>
+            lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
+            eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim
+            ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+            aliquip ex ea commodo consequat. Duis aute irure dolor in
+            reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
+            pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
+            culpa qui officia deserunt mollit anim id est laborum.
+          </div>
+        </div>
+      );
+    },
+  ],
+  parameters: {
+    layout: "padded",
+  },
+};
+
+export const ScrollLockBodyStyleTest: StoryObj = {
+  render: ScrollLockComponent,
+  decorators: [
     (story) => (
       <div
         style={{
+          height: "200vh",
+          width: "200vw",
           background: "var(--ax-bg-softA)",
         }}
       >
-        <h2>Resize to go from no scroll to scroll</h2>
+        <h2>200vh and 200vw container</h2>
         <div>{story()}</div>
-        <div style={{ width: 400, resize: "horizontal", overflow: "auto" }}>
-          lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-          eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
-          minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-          aliquip ex ea commodo consequat. Duis aute irure dolor in
-          reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-          pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-          culpa qui officia deserunt mollit anim id est laborum.
-        </div>
       </div>
     ),
+    (story) => {
+      useEffect(() => {
+        document.body.style.marginBlock = "6rem";
+      }, []);
+
+      return (
+        <>
+          <div
+            style={{
+              top: "4rem",
+              position: "absolute",
+              width: "100vw",
+              height: "4rem",
+              background: "red",
+            }}
+          />
+          {story()}
+        </>
+      );
+    },
   ],
   parameters: {
     layout: "padded",
@@ -84,7 +129,16 @@ function ScrollLockComponent() {
 
   return (
     <div>
-      <div>
+      <div style={{ position: "relative" }}>
+        <p>
+          lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
+          eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
+          minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+          aliquip ex ea commodo consequat. Duis aute irure dolor in
+          reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
+          pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
+          culpa qui officia deserunt mollit anim id est laborum.
+        </p>
         <button onClick={() => setEnabled(true)}>Enable</button>
         <button onClick={() => setEnabled(false)}>Disable</button>
       </div>