+++ aksel.nav.no/website/.storybook/main.ts
@@ -1,7 +1,7 @@
 import type { StorybookConfig } from "@storybook/nextjs";
-import { readFileSync } from "fs";
+import { readFileSync } from "node:fs";
 import { createRequire } from "node:module";
-import { dirname, join, resolve } from "path";
+import { dirname, join, resolve } from "node:path";
 import { loadCsf } from "storybook/internal/csf-tools";
 import TsconfigPathsPlugin from "tsconfig-paths-webpack-plugin";