+++ aksel.nav.no/website/scripts/update-examples/parts/filter-code.ts
@@ -1,7 +1,7 @@
 import { namedTypes } from "ast-types";
-import fs from "fs";
 import jscodeshift, { ASTPath, Collection } from "jscodeshift";
-import path from "path";
+import fs from "node:fs";
+import path from "node:path";
 // eslint-disable-next-line import/default
 import prettier from "prettier";